Hide impl. details of tracking of unsupported commands
